Add support for noise model and level 1 data to local sampler
This change passes through the `simulator.noise_model` option to the
`BackendSamplerV2` or `BackendEstimatorV2` as a `noise_model` option if
the primitive supports the `noise_model` option (the primitives in the
current release of Qiskit, 1.2.4, do not but there is an open pull
request to add support).

Additionally, this change translates the `execution.meas_type` option
into `meas_level` and `meas_return` options if the `BackendSamplerV2`
supports them. This change allows support for level 1 data in local
testing mode, where otherwise the default is only to return classified
data.